+++ CHANGELOG.rst
@@ -36,6 +36,7 @@ TODO: integration test with pandapower (see `pandapower/contingency/contingency.
 - [IMPROVED] Removing the "ChooseSolver" API and replacing it with the `SolverRegistry`
 - [IMPROVED] clean separation between lightsim2grid_core the main library that could be
   used from cpp and the python bindings.
+- [IMPROVED] removed the ".values" and replace them by ".to_numpy()" in pandapower converter.
 
 [0.13.1]  2026-04-21
 --------------------